refactor(desktop): move session list layout helpers
